Move profiling functionality out of scheduler
