+++ src/cocotb/regression.py
@@ -57,6 +57,7 @@ from typing import (
 )
 
 import cocotb
+import cocotb._profiling
 import cocotb._scheduler
 import cocotb._write_scheduler
 from cocotb import _ANSI, simulator
@@ -476,6 +477,7 @@ class RegressionManager:
         simulator.stop_simulator()
         cocotb._stop_user_coverage()
         cocotb._stop_library_coverage()
+        cocotb._profiling.finalize()
 
     def _test_complete(self) -> None:
         """Callback given to the scheduler, to be called when the current test completes.
